OeuvreGuard/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
psycopg2-binary
pydantic
sentence-transformers
simhash
faiss-cpu
numpy
scipy
librosa
pydub
python-magic
python-docx
pdfminer.six
reportlab
python-dotenv
requests
